feat:rearrange DOM structure of MainIndex,render by regular size pattern
